rename: utils 패키지 -> exception 패키지로 이름 변경
