+++ src/main/java/lotto/model/WinningLotto.java
@@ -1,8 +1,8 @@
 package lotto.model;
 
-import static lotto.utils.LottoException.DUPLICATED_BONUS_NUMBER;
-import static lotto.utils.LottoException.LESS_THAN_MINIMUM_LOTTO_NUMBER;
-import static lotto.utils.LottoException.MORE_THAN_MAXIMUM_LOTTO_NUMBER;
+import static lotto.exception.LottoException.DUPLICATED_BONUS_NUMBER;
+import static lotto.exception.LottoException.LESS_THAN_MINIMUM_LOTTO_NUMBER;
+import static lotto.exception.LottoException.MORE_THAN_MAXIMUM_LOTTO_NUMBER;
 
 public class WinningLotto {
     private static final int MINIMUM_LOTTO_NUMBER = 1;
