+++ src/test/java/lotto/model/WinningLottoTest.java
@@ -1,8 +1,8 @@
 package lotto.model;
 
-import static lotto.utils.LottoException.DUPLICATED_BONUS_NUMBER;
-import static lotto.utils.LottoException.LESS_THAN_MINIMUM_LOTTO_NUMBER;
-import static lotto.utils.LottoException.MORE_THAN_MAXIMUM_LOTTO_NUMBER;
+import static lotto.exception.LottoException.DUPLICATED_BONUS_NUMBER;
+import static lotto.exception.LottoException.LESS_THAN_MINIMUM_LOTTO_NUMBER;
+import static lotto.exception.LottoException.MORE_THAN_MAXIMUM_LOTTO_NUMBER;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
